Move implementation to a Mixin
- Remove the ES6 class
- Move implementation to a Mixin, that can be used in either the service
  or a standalone object
- Closes #106
