+++ addon/services/ajax.js
@@ -1,6 +1,5 @@
 import Ember from 'ember';
-import AjaxRequest from '../ajax-request';
-import mixinify from 'ember-mixinify-class';
+import AjaxRequestMixin from '../mixins/ajax-request';
 
 const { Service } = Ember;
 
@@ -63,4 +62,4 @@ const { Service } = Ember;
  * ```
  * @public
  */
-export default Service.extend(mixinify(AjaxRequest));
+export default Service.extend(AjaxRequestMixin);
